Folder structure fix + naming fix
