+++ finscheduler-web/src/features/dashboard/Dashboard.tsx
@@ -1,4 +1,4 @@
-import Layout from "../Main/subcomponents/Layout.tsx";
+import Layout from "../main/subcomponents/Layout.tsx";
 
 export default function Dashboard() {
     return (<Layout headerName={"Дашборды"}>
