+++ finscheduler-web/src/features/items/Items.tsx
@@ -1,17 +1,17 @@
-import DataTable, {type TableColumn} from "../dataTable/DataTable.tsx";
+import DataTable, {type TableColumn} from "../../components/dataTable/DataTable.tsx";
 import {Badge, Box, Button, Flex, Input, Text, Spinner} from "@chakra-ui/react";
 import {SearchIcon} from "lucide-react";
 import {useState, useEffect, useCallback} from "react";
-import type {TagDto, TagFilter} from "../../../api/types.ts";
-import {toaster} from "../../ui/toaster.tsx";
-import Layout from "../Main/subcomponents/Layout.tsx";
-import TagModal from "./subcomponents/TagModal.tsx";
-import TagsService from "../../../api/tags.ts";
+import type {ItemDto, ItemFilter, ItemModification} from "../../api/types.ts";
+import ItemModal from "./subcomponents/ItemModal.tsx";
+import {toaster} from "../../components/ui/toaster.tsx";
+import Layout from "../main/subcomponents/Layout.tsx";
+import ItemsService from "../../api/items.ts";
 
-const tagsService = new TagsService();
+const itemsService = new ItemsService();
 
-export default function Tags() {
-    const [tags, setTags] = useState<TagDto[]>([]);
+export default function Items() {
+    const [items, setItems] = useState<ItemDto[]>([]);
     const [total, setTotal] = useState<number>(0);
     const [loading, setLoading] = useState<boolean>(true);
     const [error, setError] = useState<string | null>(null);
@@ -19,16 +19,18 @@ export default function Tags() {
     const [pageSize, setPageSize] = useState<number>(10);
     const [selectedRows, setSelectedRows] = useState<Set<string>>(new Set());
     const [isModalOpen, setIsModalOpen] = useState(false);
-    const [editingTag, setEditingTag] = useState<TagDto | null>(null);
+    const [editingItem, setEditingItem] = useState<ItemDto | null>(null);
     const [modalMode, setModalMode] = useState<'create' | 'edit'>('create');
     const [searchTerm, setSearchTerm] = useState('');
     const [statusFilter, setStatusFilter] = useState<'All' | 'Active' | 'Inactive'>('Active');
+    const [priceFrom, setPriceFrom] = useState<string>('');
+    const [priceTo, setPriceTo] = useState<string>('');
 
-    const itemColumns: TableColumn<TagDto>[] = [
+    const itemColumns: TableColumn<ItemDto>[] = [
         {
             header: 'Название',
             key: 'name',
-            render: (row: TagDto) => (
+            render: (row: ItemDto) => (
                 <Text fontWeight="semibold" color="neon.blue">
                     {row.name || '-'}
                 </Text>
@@ -36,9 +38,34 @@ export default function Tags() {
             headerProps: {textAlign: 'left'},
         },
         {
+            header: 'Цена',
+            key: 'price',
+            render: (row: ItemDto) => (
+                <Text color="neon.blue" fontWeight="medium">
+                    {row.price !== undefined ? `₽${row.price.toLocaleString(undefined, {
+                        minimumFractionDigits: 2,
+                        maximumFractionDigits: 2
+                    })}` : '-'}
+                </Text>
+            ),
+            headerProps: {textAlign: 'right'},
+            cellProps: {justifyContent: 'flex-start'}
+        },
+        {
+            header: 'Кэшбэк (%)',
+            key: 'cashback',
+            render: (row: ItemDto) => (
+                <Text color={getCashbackColor(row.cashback)} fontWeight="bold">
+                    {row.cashback !== undefined ? `${row.cashback}%` : '-'}
+                </Text>
+            ),
+            headerProps: {textAlign: 'right'},
+            cellProps: {justifyContent: 'flex-start'}
+        },
+        {
             header: 'Статус',
             key: 'isActive',
-            render: (row: TagDto) => (
+            render: (row: ItemDto) => (
                 <Badge
                     fontSize="sm"
                     px={3}
@@ -52,13 +79,33 @@ export default function Tags() {
             ),
             headerProps: {textAlign: 'left'}
         },
+        {
+            header: 'Создан',
+            key: 'createdAt',
+            render: (row: ItemDto) => (
+                <Text color="neon.blue" fontSize="sm">
+                    {row.createdAt ? new Date(row.createdAt).toLocaleDateString('ru-RU') : '-'}
+                </Text>
+            ),
+            headerProps: {textAlign: 'left'}
+        },
     ];
 
-    const loadTags = useCallback(async () => {
+    const getCashbackColor = (cashback: number | undefined) => {
+        let color = "neon.pink";
+
+        if (cashback) {
+            color = cashback > 1 ? "neon.green" : "neon.yellow";
+        }
+
+        return color
+    }
+
+    const loadItems = useCallback(async () => {
         setLoading(true);
         setError(null);
         try {
-            const filter: TagFilter = {
+            const filter: ItemFilter = {
                 page: page - 1,
                 pageSize: pageSize,
             };
@@ -71,73 +118,108 @@ export default function Tags() {
                 filter.isActive = statusFilter === 'Active';
             }
 
-            const result = await tagsService.getTags(filter);
-            setTags(result.data);
+            if (priceFrom) {
+                const price = parseFloat(priceFrom);
+                if (!isNaN(price)) {
+                    filter.priceFrom = price;
+                }
+            }
+
+            if (priceTo) {
+                const price = parseFloat(priceTo);
+                if (!isNaN(price)) {
+                    filter.priceTo = price;
+                }
+            }
+
+            const result = await itemsService.getItems(filter);
+            setItems(result.data);
             setTotal(result.count);
         } catch (err) {
             setError(err instanceof Error ? err.message : 'Ошибка загрузки данных');
-            console.error('Failed to load tags:', err);
+            console.error('Failed to load items:', err);
         } finally {
             setLoading(false);
         }
-    }, [page, pageSize, searchTerm, statusFilter]);
+    }, [page, pageSize, searchTerm, statusFilter, priceFrom, priceTo]);
 
     useEffect(() => {
-        loadTags();
-    }, [loadTags]);
+        loadItems();
+    }, [loadItems]);
 
     const handleReset = () => {
         setSearchTerm('');
         setStatusFilter('All');
+        setPriceFrom('');
+        setPriceTo('');
         setPage(1);
     }
 
     const handleOpenAddModal = () => {
-        setEditingTag(null);
+        setEditingItem(null);
         setModalMode('create');
         setIsModalOpen(true);
     };
 
-    const handleOpenEditModal = (tag: TagDto) => {
-        setEditingTag(tag);
+    const handleOpenEditModal = (item: ItemDto) => {
+        setEditingItem(item);
         setModalMode('edit');
         setIsModalOpen(true);
     };
 
     const handleCloseModal = () => {
         setIsModalOpen(false);
-        setEditingTag(null);
+        setEditingItem(null);
         setSelectedRows(new Set());
     };
 
-    const handleSaveItem = async (tagData: Omit<TagDto, 'id'>) => {
+    const handleSaveItem = async (itemData: ItemModification) => {
         try {
             if (modalMode === 'create') {
-                await tagsService.createTag(tagData);
+                await itemsService.createItem(itemData);
                 toaster.create({
                     title: 'Успешно',
                     description: 'Элемент успешно добавлен',
                     type: 'success',
                 });
-            } else if (modalMode === 'edit' && editingTag?.id) {
-                await tagsService.updateTag(editingTag.id, tagData);
+            } else if (modalMode === 'edit' && editingItem?.id) {
+                await itemsService.updateItem(editingItem.id, itemData);
                 toaster.create({
                     title: 'Успешно',
                     description: 'Элемент успешно обновлен',
                     type: 'success',
                 });
             } else {
-                console.error('Cannot save: missing id for edit mode', {modalMode, editingItem: editingTag});
+                console.error('Cannot save: missing id for edit mode', {modalMode, editingItem});
                 throw new Error('Не удалось определить режим сохранения');
             }
-            await loadTags();
+            await loadItems();
         } catch (err) {
             console.error('Error saving item:', err);
             throw err;
         }
     };
 
-    const getRowId = (row: TagDto): string => {
+    const handleDeleteItems = async (ids: string[]) => {
+        try {
+            await Promise.all(ids.map(id => itemsService.deleteItem(id)));
+            toaster.create({
+                title: 'Успешно',
+                description: `Удалено элементов: ${ids.length}`,
+                type: 'success',
+            });
+            setSelectedRows(new Set());
+            await loadItems();
+        } catch (err) {
+            toaster.create({
+                title: 'Ошибка',
+                description: err instanceof Error ? err.message : 'Не удалось удалить элементы',
+                type: 'error',
+            });
+        }
+    };
+
+    const getRowId = (row: ItemDto): string => {
         return row.id ?? '';
     };
 
@@ -147,7 +229,7 @@ export default function Tags() {
 
     const statusOptions: Array<'All' | 'Active' | 'Inactive'> = ['All', 'Active', 'Inactive'];
 
-    return (<Layout headerName={"Тэги"}>
+    return (<Layout headerName={"Виды расходов"}>
         <Flex direction="column" width="100%">
 
             <Flex
@@ -179,7 +261,7 @@ export default function Tags() {
                         onKeyDown={(e) => {
                             if (e.key === 'Enter') {
                                 setPage(1);
-                                loadTags();
+                                loadItems();
                             }
                         }}
                         pl="10"
@@ -233,6 +315,44 @@ export default function Tags() {
                     </Flex>
                 </Box>
 
+                <Box {...filterWidthProps}>
+                    <Input
+                        placeholder="Цена от"
+                        type="number"
+                        value={priceFrom}
+                        onChange={(e) => setPriceFrom(e.target.value)}
+                        onKeyDown={(e) => {
+                            if (e.key === 'Enter') {
+                                setPage(1);
+                                loadItems();
+                            }
+                        }}
+                        bg="bg.layer1"
+                        borderColor="glass.border"
+                        color="neon.blue"
+                        _placeholder={{color: 'textMuted'}}
+                    />
+                </Box>
+
+                <Box {...filterWidthProps}>
+                    <Input
+                        placeholder="Цена до"
+                        type="number"
+                        value={priceTo}
+                        onChange={(e) => setPriceTo(e.target.value)}
+                        onKeyDown={(e) => {
+                            if (e.key === 'Enter') {
+                                setPage(1);
+                                loadItems();
+                            }
+                        }}
+                        bg="bg.layer1"
+                        borderColor="glass.border"
+                        color="neon.blue"
+                        _placeholder={{color: 'textMuted'}}
+                    />
+                </Box>
+
                 <Button
                     {...filterWidthProps}
                     onClick={handleReset}
@@ -258,7 +378,7 @@ export default function Tags() {
             ) : (
                 <>
                     <DataTable
-                        data={tags}
+                        data={items}
                         columns={itemColumns}
                         total={total}
                         page={page}
@@ -273,14 +393,15 @@ export default function Tags() {
                         onSelectionChange={setSelectedRows}
                         onAdd={handleOpenAddModal}
                         onEdit={handleOpenEditModal}
+                        onDelete={handleDeleteItems}
                         getRowId={getRowId}
                     />
-                    <TagModal
-                        key={`${modalMode}-${editingTag?.id || 'new'}`}
+                    <ItemModal
+                        key={`${modalMode}-${editingItem?.id || 'new'}`}
                         isOpen={isModalOpen}
                         onClose={handleCloseModal}
                         onSave={handleSaveItem}
-                        item={editingTag}
+                        item={editingItem}
                         mode={modalMode}
                     />
                 </>
